Move Google Pay sheet launching into GooglePayContent.
COSDK-1309
